refactor to use new errno path
